requirements.txt
streamlit==1.23.1
tensorflow==2.9.0
protobuf>=3.20,<3.21
pillow
numpy==1.21.6
opencv-python
